Fix the header inclusion problem in Linux and change parameters.

+++ CipherText.h
@@ -5,6 +5,10 @@
 #include <smmintrin.h>
 #endif
 
+#include <stdint.h>
+#include <stdio.h>
+#include <type.h>
+
 #include "ZqMatrix.h"
 
 class CipherText : public ZqMatrix 
@@ -68,7 +72,7 @@ ZqMatrix CipherText::operator*(const ZqMatrix& theMat)
 			uint64_t* b = packedB+j*packedNum;
                         int s = 0;
 			for(int k=0; k<packedNum; k++) {
-				int t = packedA[k] & packedB[k];
+				uint64 t = packedA[k] & packedB[k];
 #ifdef POPCNT_OPTIMIZATION
 				s += _mm_popcnt_u64(t);
 #else
